Function may not have return
It is hard to define value of `for`, so I decided to make it statement.
Then we need to care about when `for` is a last statement of a function.
In this case, with this patch, `return` is just omitted from function
and thus it returns `undefined`. (Or you can say "value of `for` is
`undefined` in RumCoke.)